refactor: split SeasonArchiveView into SeasonArchiveIndexView + SeasonDetailView (1.12)
Extract shared helpers into SeasonArchiveRenderHelpers trait (esc, renderPlayerName,
renderTeamCell, renderStyles) — bodies verbatim, single source of truth.

New types:
- SeasonArchiveRenderHelpers (trait) — 4 shared private helpers
- SeasonArchiveIndexView + SeasonArchiveIndexViewInterface — index page rendering
- SeasonDetailView + SeasonDetailViewInterface — detail page rendering (ROUND_NAMES +
  all 12 detail-only private methods moved verbatim)

Deleted:
- SeasonArchiveView (553-LOC hot-file)
- SeasonArchiveViewInterface
- SeasonArchiveViewTest (split into SeasonArchiveIndexViewTest + SeasonDetailViewTest)

Rewired: modules/SeasonArchive/index.php — 3 routing paths preserved byte-identically;
$indexView / $detailView replace $view; no-adr marker in PR body covers prose references
to refactor-flag/adr-check gate scripts.

All 4 golden-master snapshots byte-match post-split; full suite green (5962 tests)
