+++ ibl5/classes/SeasonArchive/Contracts/SeasonArchiveIndexViewInterface.php
@@ -5,17 +5,13 @@ declare(strict_types=1);
 namespace SeasonArchive\Contracts;
 
 /**
- * SeasonArchiveViewInterface - Contract for season archive HTML rendering
- *
- * Defines methods for generating HTML output for the season archive
- * index page and individual season detail pages.
+ * SeasonArchiveIndexViewInterface - Contract for season archive index page rendering
  *
  * @phpstan-import-type SeasonSummary from SeasonArchiveServiceInterface
- * @phpstan-import-type SeasonDetail from SeasonArchiveServiceInterface
  *
- * @see \SeasonArchive\SeasonArchiveView For the concrete implementation
+ * @see \SeasonArchive\SeasonArchiveIndexView For the concrete implementation
  */
-interface SeasonArchiveViewInterface
+interface SeasonArchiveIndexViewInterface
 {
     /**
      * Render the season archive index page
@@ -34,14 +30,4 @@ interface SeasonArchiveViewInterface
         array $playerIds = [],
         array $teamIds = []
     ): string;
-
-    /**
-     * Render a single season detail page
-     *
-     * Displays all awards, standings, bracket, and roster data for a season.
-     *
-     * @param SeasonDetail $seasonData Full season data
-     * @return string HTML output
-     */
-    public function renderSeasonDetail(array $seasonData): string;
 }
