+++ ibl5/classes/SeasonArchive/SeasonDetailView.php
@@ -4,26 +4,24 @@ declare(strict_types=1);
 
 namespace SeasonArchive;
 
-use Player\PlayerImageHelper;
 use SeasonArchive\Contracts\SeasonArchiveServiceInterface;
-use SeasonArchive\Contracts\SeasonArchiveViewInterface;
-use UI\TeamCellHelper;
-use Security\HtmlSanitizer;
+use SeasonArchive\Contracts\SeasonDetailViewInterface;
 
 /**
- * SeasonArchiveView - HTML rendering for season archive pages
+ * SeasonDetailView - HTML rendering for the season detail page
  *
- * Generates HTML for the index listing and individual season detail pages.
+ * Generates all ~12 sections for an individual season detail page.
  * All dynamic output is sanitized via HtmlSanitizer::safeHtmlOutput().
  *
- * @phpstan-import-type SeasonSummary from SeasonArchiveServiceInterface
  * @phpstan-import-type SeasonDetail from SeasonArchiveServiceInterface
  * @phpstan-import-type PlayoffSeries from SeasonArchiveServiceInterface
  *
- * @see SeasonArchiveViewInterface For the interface contract
+ * @see SeasonDetailViewInterface For the interface contract
  */
-class SeasonArchiveView implements SeasonArchiveViewInterface
+class SeasonDetailView implements SeasonDetailViewInterface
 {
+    use SeasonArchiveRenderHelpers;
+
     private const ROUND_NAMES = [
         1 => 'First Round',
         2 => 'Conference Semifinals',
@@ -32,48 +30,7 @@ class SeasonArchiveView implements SeasonArchiveViewInterface
     ];
 
     /**
-     * @see SeasonArchiveViewInterface::renderIndex()
-     *
-     * @param list<SeasonSummary> $seasons
-     * @param array<string, array{color1: string, color2: string, teamid: int}> $teamColors
-     * @param array<string, int> $playerIds
-     * @param array<string, int> $teamIds
-     */
-    public function renderIndex(
-        array $seasons,
-        array $teamColors = [],
-        array $playerIds = [],
-        array $teamIds = []
-    ): string {
-        $html = '';
-        if ($teamColors !== [] || $playerIds !== []) {
-            $html .= $this->renderStyles();
-        }
-        $html .= '<h2 class="ibl-title">IBL Season Archive</h2>';
-        $html .= '<table class="sortable ibl-data-table season-archive-index-table" data-no-responsive>';
-        $html .= '<thead><tr><th>Season</th><th>HEAT Champion</th><th>IBL Champion</th><th>MVP</th></tr></thead>';
-        $html .= '<tbody>';
-
-        foreach ($seasons as $season) {
-            /** @var array{year: int, label: string, iblChampion: string, heatChampion: string, mvp: string} $season */
-            $year = $season['year'];
-            $label = self::esc($season['label']);
-
-            $html .= '<tr>';
-            $html .= '<td><a href="modules.php?name=SeasonArchive&amp;year=' . $year . '">' . $label . '</a></td>';
-            $html .= self::renderTeamCell($season['heatChampion'], $teamColors, $year);
-            $html .= self::renderTeamCell($season['iblChampion'], $teamColors, $year);
-            $html .= '<td>' . self::renderPlayerName($season['mvp'], $playerIds) . '</td>';
-            $html .= '</tr>';
-        }
-
-        $html .= '</tbody></table>';
-
-        return $html;
-    }
-
-    /**
-     * @see SeasonArchiveViewInterface::renderSeasonDetail()
+     * @see SeasonDetailViewInterface::renderSeasonDetail()
      *
      * @param SeasonDetail $seasonData
      */
@@ -137,64 +94,6 @@ class SeasonArchiveView implements SeasonArchiveViewInterface
         $html .= $this->renderRosters($championRosters, $allStarRosters, $allStarCoaches, $playerIds, $teamColors, $year, $iblChampion, $iblChampionCoach);
 
         return $html;
-    }
-
-    /**
-     * Escape a string value for safe HTML output
-     *
-     * Wraps HtmlSanitizer::safeHtmlOutput() with a string return type cast
-     * to satisfy PHPStan strict type checking in string concatenation contexts.
-     */
-    private static function esc(string $value): string
-    {
-        /** @var string */
-        return HtmlSanitizer::safeHtmlOutput($value);
-    }
-
-    /**
-     * Render a player name with optional photo thumbnail and link
-     *
-     * @param array<string, int> $playerIds Map of player name => pid
-     */
-    private static function renderPlayerName(string $name, array $playerIds): string
-    {
-        if ($name === '') {
-            return '';
-        }
-
-        $pid = $playerIds[$name] ?? null;
-        if ($pid !== null) {
-            return '<span class="ibl-player-cell">'
-                . PlayerImageHelper::renderPlayerLink($pid, $name)
-                . '</span>';
-        }
-
-        return self::esc($name);
-    }
-
-    /**
-     * Render a team cell with colored background, logo, and link
-     *
-     * @param array<string, array{color1: string, color2: string, teamid: int}> $teamColors
-     */
-    private static function renderTeamCell(string $teamName, array $teamColors, int $year): string
-    {
-        $colors = $teamColors[$teamName] ?? null;
-        if ($colors !== null) {
-            $teamid = $colors['teamid'];
-            $yearUrl = TeamCellHelper::teamPageUrl($teamid, $year);
-            return TeamCellHelper::renderTeamCell($teamid, $teamName, $colors['color1'], $colors['color2'], '', $yearUrl);
-        }
-
-        return '<td>' . self::esc($teamName) . '</td>';
-    }
-
-    /**
-     * Styles are now centralized in design/components/season-archive.css.
-     */
-    private function renderStyles(): string
-    {
-        return '';
     }
 
     /**
